requirements.txt
# LLM SDKs
anthropic>=0.40.0
openai>=1.0.0

# MCP SDK (Anthropic's official Python SDK)
mcp>=1.0.0

# LangGraph orchestration
langgraph>=0.2.0
langchain-core>=0.3.0

# Database
asyncpg>=0.29.0
duckdb>=0.10.0
pgvector>=0.3.0

# Redis
redis>=5.0.0

# HTTP server and client
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-multipart>=0.0.9

# Observability
langfuse>=2.0.0

# Data processing (used inside sandbox-executed transformation scripts only)
pandas>=2.2.0
pyarrow>=14.0.0
numpy>=1.26.0

# Config and validation
pydantic>=2.7.0
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
